✓☺ typed all variables except the variable named child.
There is a variable named child in the onSave2 file. This variable is
assigned to a child process that is revealed by the return value of the
spawn function.

The value is typed by the @types/node type library. This type definition
is very large. And I must find a way to retrieve it using its name.

I don't want to copy and paste it from the source file because that
would ruin the code.

There must be another way.